chore: cleanup tests and refactor pkg structure.
