+++ pkg/probes-handler/probes_handler.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/sirupsen/logrus"
+	"github.com/arquillian/ike-prow-plugins/pkg/log"
 )
 
 // Probe defines the json elements for the probesHandler response body content
@@ -15,12 +15,12 @@ type Probe struct {
 }
 
 // NewProbesHandler registers liveliness and readinesss probes for /version endpoint
-func NewProbesHandler(log *logrus.Entry) http.Handler {
+func NewProbesHandler(log log.Logger) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 
 		serverError := func(action string, err error) {
-			log.WithError(err).Errorf("Error %s.", action)
-			msg := fmt.Sprintf("500 Internal server error %s: %v", action, err)
+			log.Errorf("Probe handler failed when %s: %v.", action, err)
+			msg := fmt.Sprintf("Internal server error %d %s: %v", http.StatusInternalServerError, action, err)
 			http.Error(w, msg, http.StatusInternalServerError)
 		}
 
@@ -29,10 +29,10 @@ func NewProbesHandler(log *logrus.Entry) http.Handler {
 
 		content, err := getJSONContent()
 		if err != nil {
-			serverError("marshaling response body: %v", err)
+			serverError("marshaling response body", err)
 		}
 		if _, err := w.Write(content); err != nil {
-			serverError("writing response body: %v", err)
+			serverError("writing response body", err)
 		}
 	})
 }
